Moved all new Exceptions to sub-packages
 Changes to be committed:
	renamed:    src/main/java/com/ardaslegends/service/exceptions/AlreadyExistingException.java -> src/main/java/com/ardaslegends/service/alreadyexisting/AlreadyExistingException.java
	renamed:    src/main/java/com/ardaslegends/service/exceptions/LogicException.java -> src/main/java/com/ardaslegends/service/exceptions/logic/LogicException.java
